# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Wow Libre Core API
  description: API REST para la gestión de usuarios y autenticación del sistema Wow Libre
  version: 1.0.0
paths:
  /api/account/create:
    post:
      tags:
        - User Management
      summary: Create new user account
      description: Creates a new user account and returns JWT authentication tokens
      operationId: createUser
      parameters:
        - name: X-Transaction-ID
          in: header
          required: false
          schema:
            type: string
            example: "txn-123456"
        - name: Accept-Language
          in: header
          required: false
          schema:
            type: string
            enum: [en, es, pt]
            example: "en"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDto"
            examples:
              userCreation:
                summary: Complete user registration data
                value:
                  country: "US"
                  date_of_birth: "1990-01-15"
                  first_name: "John"
                  last_name: "Doe"
                  cell_phone: "+1234567890"
                  email: "john.doe@example.com"
                  password: "SecurePass123!"
                  language: "en"
                  token: "03AGdBq25..."
      responses:
        "201":
          description: User account created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtResponse"
              examples:
                success:
                  summary: Successful user creation
                  value:
                    data:
                      id: 123
                      jwt: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      expiration_date: "2024-01-15T10:30:00Z"
                      avatar_url: "https://example.com/avatar.jpg"
                      language: "en"
                      pending_validation: true
                      isAdmin: false
                    transaction_id: "txn-123456"
                    status: "CREATED"
                    message: "User created successfully"
        "400":
          description: Bad request - Invalid input data
        "409":
          description: Conflict - User already exists
        "500":
          description: Internal server error

components:
  schemas:
    UserDto:
      type: object
      required:
        - country
        - date_of_birth
        - first_name
        - last_name
        - email
        - password
        - language
        - token
      properties:
        country:
          type: string
          minLength: 2
          maxLength: 30
          example: "US"
        date_of_birth:
          type: string
          format: date
          example: "1990-01-15"
        first_name:
          type: string
          minLength: 3
          maxLength: 30
          example: "John"
        last_name:
          type: string
          minLength: 3
          maxLength: 30
          example: "Doe"
        cell_phone:
          type: string
          example: "+1234567890"
        email:
          type: string
          format: email
          minLength: 5
          maxLength: 50
          example: "john.doe@example.com"
        password:
          type: string
          maxLength: 30
          example: "SecurePass123!"
        language:
          type: string
          enum: [en, es, pt]
          example: "en"
        token:
          type: string
          example: "03AGdBq25..."

    JwtResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/JwtDto"
        transaction_id:
          type: string
        status:
          type: string
        message:
          type: string

    JwtDto:
      type: object
      properties:
        id:
          type: integer
        jwt:
          type: string
        refresh_token:
          type: string
        expiration_date:
          type: string
          format: date-time
        avatar_url:
          type: string
        language:
          type: string
        pending_validation:
          type: boolean
        isAdmin:
          type: boolean